fix:알람템플릿 수정 , 채팅 상세기능 조정
