+++ Linkee-api/src/main/java/com/linkee/linkeeapi/alarm/command/domain/aggregate/entity/AlarmBox.java
@@ -27,28 +27,9 @@ public class AlarmBox extends BaseTimeEntity {
     @JoinColumn(name = "user_id", referencedColumnName = "user_id")
     private User user;
 
-
-
-
     public void checkedAlarm(){
         this.isChecked = Status.Y;
     }
 
-    @NoArgsConstructor
-    @AllArgsConstructor
-    @Builder
-    @ToString
-    @Getter
-    @Entity
-    @Table(name = "tb_alarm_template")
-    public static class AlarmTemplate extends BaseTimeEntity {
-        @Id
-        @GeneratedValue(strategy = GenerationType.IDENTITY)
-        private Long templateId;
-        private String templateContent;
 
-        public void modifyTemplateContent(String content) {
-            this.templateContent = content;
-        }
-    }
 }
