+++ Linkee-api/src/main/java/com/linkee/linkeeapi/alarm/command/instructure/repository/AlarmTemplateRepository.java
@@ -1,7 +1,7 @@
 package com.linkee.linkeeapi.alarm.command.instructure.repository;
 
-import com.linkee.linkeeapi.alarm.command.domain.aggregate.entity.AlarmBox;
+import com.linkee.linkeeapi.alarm.command.domain.aggregate.entity.AlarmTemplate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlarmTemplateRepository extends JpaRepository<AlarmBox.AlarmTemplate,Long> {
+public interface AlarmTemplateRepository extends JpaRepository<AlarmTemplate,Long> {
 }
