+++ Linkee-api/src/main/resources/templates/chat.html
@@ -6,17 +6,98 @@
     <script src="https://cdn.jsdelivr.net/npm/sockjs-client@1.6.1/dist/sockjs.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/stompjs@2.3.3/lib/stomp.min.js"></script>
     <style>
-        body { font-family: Arial, sans-serif; }
-        #rooms { list-style: none; padding: 0; margin: 0; width: 300px; }
-        #rooms li { padding: 8px; border: 1px solid #ccc; margin-bottom: 4px; cursor: pointer; display: flex; justify-content: space-between; }
-        #rooms li:hover { background-color: #f0f0f0; }
-        #rooms li.active { background-color: #d0f0d0; }
-        #chat { border: 1px solid #ccc; height: 400px; overflow-y: scroll; padding: 10px; margin-top: 10px; width: 600px; }
-        #message { width: 500px; }
+        body { font-family: Arial, sans-serif; margin: 20px; background-color: #f4f4f4; }
+        h2 { margin-bottom: 10px; }
+        input, button { font-size: 14px; }
+
+        /* 방 목록 */
+        #rooms-container {
+            width: 250px;
+            flex-shrink: 0;
+        }
+
+        #rooms { list-style: none; padding: 0; margin: 0; }
+        #rooms li {
+            padding: 8px;
+            border: 1px solid #ccc;
+            margin-bottom: 4px;
+            cursor: pointer;
+            display: flex;
+            justify-content: space-between;
+            border-radius: 4px;
+            background-color: #fff;
+        }
+        #rooms li:hover { background-color: #e0f7fa; }
+        #rooms li.active { background-color: #b2ebf2; }
+
+        #chat-container {
+            display: flex;
+            flex-direction: column;
+            height: 500px; /* 고정 높이 */
+            border: 1px solid #ccc;
+            border-radius: 6px;
+            overflow: hidden; /* 부모 넘침 방지 */
+            background-color: #fff;
+        }
+
+        #chat {
+            flex: 1; /* 부모 높이 안에서 남는 공간 채움 */
+            padding: 10px;
+            overflow-y: auto; /* 메시지 많아지면 스크롤 */
+            background-color: #fafafa;
+        }
+
+        #qna-box {
+            position: sticky;
+            top: 0;
+            background: #fff;
+            z-index: 10;
+            padding: 8px;
+            border-bottom: 1px solid #ccc;
+        }
+
+
+
+        #message-container {
+            display: flex;
+            margin-top: 10px;
+            gap: 6px;
+            align-items: center;
+        }
+
+        #message {
+            flex: 1;
+            padding: 6px;
+            border-radius: 4px;
+            border: 1px solid #ccc;
+        }
+
         .system { color: gray; font-style: italic; }
-        /* 선택 버튼 스타일 */
+
+        /* 참여자 영역 */
+        #members-container {
+            width: 150px;
+            border: 1px solid #ccc;
+            border-radius: 6px;
+            height: 200px; /* 참여자 영역 높이 줄임 */
+            overflow-y: auto; /* 참여자 많으면 스크롤 */
+            padding: 10px;
+            background-color: #fff;
+        }
+
+        #members-container h4 {
+            margin-top: 0;
+            margin-bottom: 8px;
+            font-size: 16px;
+            border-bottom: 1px solid #ccc;
+            padding-bottom: 4px;
+        }
+
+        #member-list { list-style: none; padding: 0; margin: 0; }
+
+        /* 방 타입/공개 버튼 */
         .option-btn {
-            margin: 4px;
+            margin: 4px 2px;
             padding: 6px 12px;
             border: 1px solid #999;
             background-color: #f9f9f9;
@@ -28,9 +109,15 @@
             color: white;
             font-weight: bold;
         }
+
+        /* 전체 레이아웃 */
+        #main-wrapper {
+            display: flex;
+            gap: 10px;
+            margin-top: 10px;
+        }
     </style>
 </head>
-
 <body>
 <h2>링키</h2>
 
@@ -57,13 +144,11 @@
     <button class="option-btn" id="btnPrivate" onclick="selectIsPrivate('Y')">비공개</button>
 </div>
 
-<!-- 🔹 방 코드 (비공개일 때만 표시) -->
 <div id="roomCodeContainer" style="display:none;">
     <p><strong>비공개 방 코드 입력:</strong></p>
     <input type="number" id="roomCode" placeholder="예: 1234">
 </div>
 
-<!-- 🔹 게임방 인원수 (GAME일 때만 표시) -->
 <div id="capacityContainer" style="display:none;">
     <p><strong>방 인원 수 (최대 5명):</strong></p>
     <input type="number" id="roomCapacity" min="1" max="5" value="2">
@@ -71,37 +156,41 @@
 
 <button onclick="createRoom()">방 만들기</button>
 
-<h3>방 목록</h3>
-<ul id="rooms"></ul>
+<hr>
 
-<h3>채팅</h3>
-<div style="display: flex; gap: 10px;">
-    <!-- 채팅 메시지 영역 -->
-    <div id="chat-container" style="position: relative; width: 450px; border: 1px solid #ccc; height: 400px; overflow-y: auto;">
-        <!-- QnA 영역 고정 -->
-        <div id="qna-box" style="position: sticky; top: 0; background: white; z-index: 10; padding: 5px; border-bottom: 1px solid #ccc;">
-            <p><strong>🧩 현재 문제:</strong> <span id="current-question">-</span></p>
-            <p id="answer-section"><strong>✅ 정답:</strong> <span id="current-answer"></span></p>
-            <button id="reveal-btn" onclick="revealAnswer()">정답 공개</button>
+<div id="main-wrapper">
+    <!-- 방 목록 -->
+    <div id="rooms-container">
+        <h4>💬 방 목록</h4>
+        <ul id="rooms"></ul>
+    </div>
+
+    <!-- 채팅 + QnA -->
+    <div style="flex:1; display:flex; flex-direction:column;">
+        <div id="chat-container">
+            <div id="qna-box">
+                <p><strong>🧩 현재 문제:</strong> <span id="current-question">-</span></p>
+                <p id="answer-section"><strong>✅ 정답:</strong> <span id="current-answer"></span></p>
+                <button id="reveal-btn" onclick="revealAnswer()">정답 공개</button>
+            </div>
+            <div id="chat"></div>
         </div>
 
-        <!-- 메시지 영역 -->
-        <div id="chat" style="padding: 5px;"></div>
+        <div id="message-container">
+            <input type="text" id="message" placeholder="메시지 입력">
+            <button onclick="sendMessage()">전송</button>
+            <button onclick="leaveRoom()">나가기</button>
+        </div>
+
+        <button style="margin-top: 6px;" onclick="createQuestion()">문제 등록</button>
     </div>
 
-    <!-- 참가자 리스트 영역 -->
-    <div id="members-container" style="width: 150px; border: 1px solid #ccc; height: 400px; overflow-y: auto; padding: 5px;">
-        <h4>👥 참여자</h4>
-        <ul id="member-list" style="list-style: none; padding: 0; margin: 0;"></ul>
+    <!-- 참여자 리스트 -->
+    <div id="members-container">
+        <h4 id="member-header">👥 참여자</h4>
+        <ul id="member-list"></ul>
     </div>
 </div>
-
-<input type="text" id="message" placeholder="메시지 입력">
-<button onclick="sendMessage()">전송</button>
-<button onclick="leaveRoom()">나가기</button>
-
-<!-- ✅ QnA 등록 버튼 -->
-<button onclick="createQuestion()">문제 등록</button>
 
 <script>
     let stompClient = null;
@@ -325,6 +414,12 @@
             li.textContent = m.userNickname;
             ul.appendChild(li);
         });
+        const header = document.getElementById("member-header");
+        if (selectedRoomId) {
+            header.textContent = `👥 참여자 (방 ${selectedRoomId})`;
+        } else {
+            header.textContent = `👥 참여자`;
+        }
     }
 
 
@@ -506,7 +601,11 @@
         // UI 초기화
         showMessage("방을 나갔습니다.", true);
         selectedRoomId = null;
+
         document.getElementById("chat").innerHTML = "";
+        document.getElementById("member-list").innerHTML = "";
+        document.getElementById("member-header").textContent = `👥 참여자`;
+
         highlightSelectedRoom();
         loadRooms(); // 방 목록 새로고침 (참여자 수 업데이트)
     }
